move shared utils to external package
