+++ apps/server/src/translate.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import { v2 } from '@google-cloud/translate';
 import { logDebug, log } from './logger.js';
-import * as utils from './utils.js';
+import * as utils from '@ethgoose/utils';
 import { fileURLToPath, URL } from 'url';
 const { apiKey }:GooseConfig['translate'] = JSON.parse(fs.readFileSync(fileURLToPath(new URL('../config/config.json', import.meta.url).toString()), 'utf-8')).translate;
 import validator from 'validator';
