+++ packages/utils/package.json
@@ -14,6 +14,7 @@
     "typescript": "^5.7.3"
   },
   "exports": {
+    ".": "./build/utils.js",
     "./regex": "./build/regexes.js"
   }
 }
